Changes for VS 2008 build

+++ src/drivers/lliaw/pathfinder.cpp
@@ -299,7 +299,7 @@ void Pathfinder::plan(MyCar* myc, int currentsegid)
 
 		// Compute optimal path.
 		for (int step = 128; (step /= 2) > 0;) {
-			for (int i = 100 * int(sqrt(step)); --i >= 0;) smooth(step);
+			for (int i = 100 * int(sqrt((double)step)); --i >= 0;) smooth(step);
 			interpolate(step);
 		}
 		optpathinitialized = true;
